% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winbox.R
\name{update_winbox}
\alias{update_winbox}
\title{update_winbox}
\usage{
update_winbox(
  id,
  action = c("minimize", "maximize", "fullscreen", "close", "force_close", "update"),
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Box id.}

\item{action}{Action to trigger: \code{c("minimize", "maximize", "fullscreen", "close", "force_close", "update")}.}

\item{options}{If action is update, a list of new options to configure the box, such as
\code{list(title = "new title", color = "red",
width = "500px", height = "200px", x = "center", y = "center")}.}

\item{session}{Shiny session.}
}
\description{
update_winbox
}
